Refactor Model interfaces.
Model stubs in the tests have a lot of unused methods.

Add one more layer of abstraction to meet ISP.
