+++ gradle.properties
@@ -4,3 +4,4 @@ org.gradle.jvmargs=-XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Xm
 # TODO: This is a workaround for a JDK11 bug which causes test coverage upload to fail.
 # Remove it when https://bugs.openjdk.java.net/browse/JDK-8221253 is fixed.
 systemProp.jdk.tls.client.protocols="TLSv1,TLSv1.1,TLSv1.2"
+
