+++ src/main/java/seedu/address/logic/commands/addressbookcommands/ClearCommand.java
@@ -5,19 +5,19 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
+import seedu.address.model.AddressBookModel;
 
 /**
  * Clears the address book.
  */
-public class ClearCommand extends Command {
+public class ClearCommand extends Command<AddressBookModel> {
 
     public static final String COMMAND_WORD = "clear";
     public static final String MESSAGE_SUCCESS = "Address book has been cleared!";
 
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(AddressBookModel model) {
         requireNonNull(model);
         model.setAddressBook(new AddressBook());
         return new CommandResult(MESSAGE_SUCCESS);
