+++ src/main/java/seedu/address/logic/commands/addressbookcommands/ExitCommand.java
@@ -2,19 +2,19 @@ package seedu.address.logic.commands.addressbookcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.model.Model;
+import seedu.address.model.AddressBookModel;
 
 /**
  * Terminates the program.
  */
-public class ExitCommand extends Command {
+public class ExitCommand extends Command<AddressBookModel> {
 
     public static final String COMMAND_WORD = "exit";
 
     public static final String MESSAGE_EXIT_ACKNOWLEDGEMENT = "Exiting Address Book as requested ...";
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(AddressBookModel model) {
         return new CommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT, false, true);
     }
