+++ src/main/java/seedu/address/logic/commands/addressbookcommands/FindCommand.java
@@ -5,14 +5,14 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.model.Model;
+import seedu.address.model.AddressBookModel;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 /**
  * Finds and lists all persons in address book whose name contains any of the argument keywords.
  * Keyword matching is case insensitive.
  */
-public class FindCommand extends Command {
+public class FindCommand extends Command<AddressBookModel> {
 
     public static final String COMMAND_WORD = "find";
 
@@ -28,7 +28,7 @@ public class FindCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(AddressBookModel model) {
         requireNonNull(model);
         model.updateFilteredPersonList(predicate);
         return new CommandResult(
