+++ src/main/java/seedu/address/logic/commands/schedulecommands/ScheduleSearchCommand.java
@@ -5,10 +5,10 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.model.Model;
+import seedu.address.model.ScheduleModel;
 import seedu.address.model.task.TitleContainsKeywordsPredicate;
 
-public class ScheduleSearchCommand extends Command {
+public class ScheduleSearchCommand extends Command<ScheduleModel> {
     public static final String COMMAND_WORD = "search task";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all tasks whose names contain any of "
@@ -23,7 +23,7 @@ public class ScheduleSearchCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(Model model) {
+    public CommandResult execute(ScheduleModel model) {
         requireNonNull(model);
         model.updateFilteredTaskList(predicate);
         return new CommandResult(
