+++ src/main/java/seedu/address/model/Model.java
@@ -17,10 +17,7 @@ import seedu.address.model.task.Task;
 /**
  * The API of the Model component.
  */
-public interface Model {
-    /** {@code Predicate} that always evaluate to true */
-    Predicate<Person> PREDICATE_SHOW_ALL_PERSONS = unused -> true;
-
+public interface Model extends AddressBookModel, ScheduleModel, FlashcardModel, QuizModel {
     /** {@code Predicate} that always evaluate to true */
     Predicate<Task> PREDICATE_SHOW_ALL_TASKS = unused -> true;
 
@@ -44,145 +41,4 @@ public interface Model {
      */
     void setGuiSettings(GuiSettings guiSettings);
 
-    // ADDRESSBOOK
-    /**
-     * Returns the user prefs' address book file path.
-     */
-    Path getAddressBookFilePath();
-
-    /**
-     * Sets the user prefs' address book file path.
-     */
-    void setAddressBookFilePath(Path addressBookFilePath);
-
-    /**
-     * Replaces address book data with the data in {@code addressBook}.
-     */
-    void setAddressBook(ReadOnlyAddressBook addressBook);
-
-    /** Returns the AddressBook */
-    ReadOnlyAddressBook getAddressBook();
-
-    /**
-     * Returns true if a person with the same identity as {@code person} exists in the address book.
-     */
-    boolean hasPerson(Person person);
-
-    /**
-     * Deletes the given person.
-     * The person must exist in the address book.
-     */
-    void deletePerson(Person target);
-
-    /**
-     * Adds the given person.
-     * {@code person} must not already exist in the address book.
-     */
-    void addPerson(Person person);
-
-    /**
-     * Replaces the given person {@code target} with {@code editedPerson}.
-     * {@code target} must exist in the address book.
-     * The person identity of {@code editedPerson} must not be the same as another existing person in the address book.
-     */
-    void setPerson(Person target, Person editedPerson);
-
-    /** Returns an unmodifiable view of the filtered person list */
-    ObservableList<Person> getFilteredPersonList();
-
-    /**
-     * Updates the filter of the filtered person list to filter by the given {@code predicate}.
-     * @throws NullPointerException if {@code predicate} is null.
-     */
-    void updateFilteredPersonList(Predicate<Person> predicate);
-
-    /**
-     * Returns the user prefs' schedule file path.
-     */
-    Path getScheduleFilePath();
-
-    /**
-     * Sets the user prefs' schedule file path.
-     */
-    void setScheduleFilePath(Path scheduleFilePath);
-
-    /**
-     * Replaces schedule data with the data in {@code schedule}.
-     */
-    void setSchedule(ReadOnlySchedule schedule);
-
-    /** Returns the Schedule */
-    ReadOnlySchedule getSchedule();
-
-    /**
-     * Returns true if a task with the same identity as {@code task} exists in the schedule.
-     */
-    boolean hasTask(Task task);
-
-    /**
-     * Deletes the given task.
-     * The task must exist in the schedule.
-     */
-    void deleteTask(Task target);
-
-    /**
-     * Adds the given task.
-     * {@code task} must not already exist in the task.
-     */
-    void addTask(Task task);
-
-    /**
-     * Replaces the given task {@code target} with {@code editedTask}.
-     * {@code target} must exist in the schedule.
-     * The task identity of {@code editedTask} must not be the same as another existing task in the schedule.
-     */
-    void setTask(Task target, Task editedTask);
-
-    /** Returns an unmodifiable view of the filtered task list */
-    ObservableList<Task> getFilteredTaskList();
-
-    /**
-     * Updates the filter of the filtered task list to filter by the given {@code predicate}.
-     * @throws NullPointerException if {@code predicate} is null.
-     */
-    void updateFilteredTaskList(Predicate<Task> predicate);
-
-    // FLASHCARD
-
-    void setFlashcardBank(ReadOnlyFlashcardBank flashcardBank);
-
-    ReadOnlyFlashcardBank getFlashcardBank();
-
-    boolean hasFlashcardSet(FlashcardSet flashcardSet);
-
-    void deleteFlashcardSet(FlashcardSet target);
-
-    void addFlashcardSet(FlashcardSet flashcardSet);
-
-    void setFlashcardSet(FlashcardSet target, FlashcardSet editedFlashcardSet);
-
-    ObservableList<FlashcardSet> getFlashcardSetList();
-
-    void updateFilteredFlashcardSetList(Predicate<FlashcardSet> predicate);
-
-
-
-    void addFlashcard(Flashcard flashcard, Index flashcardSetIndex);
-
-    FlashcardSet getFlashcardSet(int index); // added because quiz needs, feel free to change implementation
-
-    // QUIZ
-    Question start(Quiz quiz);
-
-    boolean hasStarted();
-
-    void tallyScore(boolean isCorrect);
-
-    Question getQuestion();
-
-    Answer getAnswer();
-
-    double stopQuiz();
-
-    String getQuizRecords(int index);
 }
