+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -3,10 +3,10 @@ package seedu.address.logic;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX;
 import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
-import static seedu.address.logic.commands.CommandTestUtil.ADDRESS_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.ADDRESS_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.PHONE_DESC_AMY;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.AMY;
