+++ src/test/java/seedu/address/logic/commands/addressbookcommands/EditCommandTest.java
@@ -2,14 +2,14 @@ package seedu.address.logic.commands.addressbookcommands;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
-import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.showPersonAtIndex;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.DESC_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.assertCommandFailure;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.assertCommandSuccess;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.showPersonAtIndex;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
 import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
@@ -18,8 +18,6 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.addressbookcommands.ClearCommand;
-import seedu.address.logic.commands.addressbookcommands.EditCommand;
 import seedu.address.logic.commands.addressbookcommands.EditCommand.EditPersonDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.FlashcardBank;
