+++ src/test/java/seedu/address/logic/parser/addressbookparsers/AddCommandParserTest.java
@@ -1,29 +1,29 @@
 package seedu.address.logic.parser.addressbookparsers;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.ADDRESS_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.ADDRESS_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_ADDRESS_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_EMAIL_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NAME_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_PHONE_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.PHONE_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_NON_EMPTY;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_WHITESPACE;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_HUSBAND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.ADDRESS_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.ADDRESS_DESC_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.EMAIL_DESC_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.INVALID_ADDRESS_DESC;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.INVALID_EMAIL_DESC;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.INVALID_NAME_DESC;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.INVALID_PHONE_DESC;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.INVALID_TAG_DESC;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.NAME_DESC_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.PHONE_DESC_AMY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.PHONE_DESC_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.PREAMBLE_NON_EMPTY;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.PREAMBLE_WHITESPACE;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.TAG_DESC_FRIEND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.TAG_DESC_HUSBAND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_TAG_FRIEND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static seedu.address.testutil.TypicalPersons.AMY;
