+++ src/test/java/seedu/address/model/person/PersonTest.java
@@ -2,11 +2,11 @@ package seedu.address.model.person;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_PHONE_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_ADDRESS_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_PHONE_BOB;
+import static seedu.address.logic.commands.commandtestutils.AddressCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalPersons.ALICE;
 import static seedu.address.testutil.TypicalPersons.BOB;
